Remove serialization package
